create and delete cells only if necessary new index page

+++ grails-app/controllers/second_project/GridController.groovy
@@ -10,63 +10,70 @@ class GridController {
         respond gridList, model: [gridList: gridList]
     }
 
-    def create(GridCommand gridCommand){
+    def create(GridCommand gridCommand) {
         gridCommand.clearErrors()
         [gridCommand: gridCommand]
     }
 
-    def save (GridCommand gridCommand){
+    def save(GridCommand gridCommand) {
         def grid = gridService.save(gridCommand)
         request.withFormat {
             form multipartForm {
                 flash.message = message(code: 'default.created.message', args: [message(code: 'grid.label', default: 'Grid'), grid.id])
                 redirect grid
             }
-            '*' {respond grid, [status: CREATED]}
+            '*' { respond grid, [status: CREATED] }
         }
     }
 
-    def show(Grid grid){
+    def show(Grid grid) {
         def cellList = Cell.findAllByGrid(grid)
 
         [grid: grid, cellList: cellList]
     }
 
 //    https://en.wikipedia.org/wiki/Conway%27s_Game_of_Life
-    def evolute (Grid grid) {
+    def evolute(Grid grid) {
+
         def cellList = Cell.findAllByGrid(grid)
 
         def simpleGrid = new int[grid.x][grid.y]
         def deleteGrid = new int[grid.x][grid.y]
         def createGrid = new int[grid.x][grid.y]
 
+        def createCells = false;
+        def deleteCells = false;
+
         cellList.each { cell ->
-            simpleGrid[cell.x-1][cell.y-1] = 1
+            simpleGrid[cell.x - 1][cell.y - 1] = 1
         }
 
-        for (def x = 0; x < grid.x; x++){
-            for (def y = 0; y < grid.y; y++){
+        for (def x = 0; x < grid.x; x++) {
+            for (def y = 0; y < grid.y; y++) {
 
                 def numberofNeigbours = 0
 
+                // infinite space
                 def xMinusOne = x - 1;
                 def yMinusOne = y - 1;
 
                 def xPlusOne = x + 1
                 def yPlusOne = y + 1
 
-                if (xMinusOne == -1){
-                    xMinusOne = grid.x-1
+                if (xMinusOne == -1) {
+                    xMinusOne = grid.x - 1
                 }
-                if (yMinusOne == -1){
-                    yMinusOne = grid.y-1
+                if (yMinusOne == -1) {
+                    yMinusOne = grid.y - 1
                 }
-                if (xPlusOne == grid.x){
+                if (xPlusOne == grid.x) {
                     xPlusOne = 0;
                 }
-                if (yPlusOne == grid.y){
+                if (yPlusOne == grid.y) {
                     yPlusOne = 0;
                 }
+
+                //calculate the number of neighbours
                 numberofNeigbours += simpleGrid[xMinusOne][yMinusOne]
                 numberofNeigbours += simpleGrid[x][yMinusOne]
                 numberofNeigbours += simpleGrid[xPlusOne][yMinusOne]
@@ -78,29 +85,33 @@ class GridController {
                 numberofNeigbours += simpleGrid[x][yPlusOne]
                 numberofNeigbours += simpleGrid[xPlusOne][yPlusOne]
 
-
-                if (simpleGrid[x][y] == 1){
-                    if (numberofNeigbours < 2){
+                //game of life algorithm
+                if (simpleGrid[x][y] == 1) {
+                    if (numberofNeigbours < 2) {
                         deleteGrid[x][y] = 1
-//                        cellService.deleteWithCoordinates(x+1,y+1,grid)
+                        deleteCells = true;
                     }
-                    if (numberofNeigbours > 3){
+                    if (numberofNeigbours > 3) {
                         deleteGrid[x][y] = 1
-//                        cellService.deleteWithCoordinates(x+1,y+1,grid)
+                        deleteCells = true;
                     }
                 } else {
-                    if (numberofNeigbours == 3){
+                    if (numberofNeigbours == 3) {
                         createGrid[x][y] = 1
-                        cellService.createWithCoordinate(x+1,y+1,grid)
+                        createCells = true;
                     }
                 }
             }
         }
 
-        cellService.deleteWithCoordinateGrid(deleteGrid, grid)
+        if (createCells) {
+            cellService.createWithCoordinateGrid(createGrid, grid)
+        }
+        if (deleteCells) {
+            cellService.deleteWithCoordinateGrid(deleteGrid, grid)
+        }
 
-
-        render(template: "game_of_life_table", model:[grid: grid, cellList: cellList])
+        render(template: "game_of_life_table", model: [grid: grid, cellList: cellList])
 
     }
